feat: exampleを整理してnode-adaptorとdiscordjs-adaptorを分離
