+++ packages/example/tsup.config.ts
@@ -1,5 +1,7 @@
 import { createTsupConfig } from "../../tsup.config";
 
 export default createTsupConfig({
+  entry: [],
   format: ["esm"],
+  target: "node18",
 });
